Pure moving of code around -- no code modification
This commit can be safely ignored when reviewing code; it
has no semantic change although it clearly does not work.
